requirements.txt
Flask==1.1.2
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib
flask_wtf
octorest